requirements.txt
affine==2.4.0
asgiref==3.6.0
aspose-pdf==23.8.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
Cartopy @ file:///C:/Users/Administrator/Downloads/Cartopy-0.20.2-cp38-cp38-win_amd64.whl#sha256=feae60ea2450e50a14caae345e94b56d4bce0bb65ac7977c80b144398abb58b5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.25.2
colorama==0.4.6
comm==0.1.3
contextily==1.3.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
descartes==1.1.0
distro==1.8.0
Django==4.1.6
executing==1.2.0
Fiona @ file:///C:/Users/Administrator/Downloads/Fiona-1.8.21-cp38-cp38-win_amd64.whl#sha256=75696223071188f5fa1c0fc5e61cefb073796365ce263c8b65026f311ba90aef
fonttools==4.39.3
fuzzywuzzy==0.18.0
GDAL @ file:///C:/Users/Administrator/Downloads/GDAL-3.4.3-cp38-cp38-win_amd64.whl#sha256=47e49a1833b7db6821bf6bad875f095bdbcba83383d9e77db557854626b55940
geocoder==0.9.1
geographiclib==2.0
geopandas==0.12.2
geoplot==0.5.1
geopy==2.3.0
idna==3.4
importlib-metadata==6.4.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jieba==0.42.1
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
Levenshtein==0.20.9
mapclassify==2.5.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mercantile==1.2.1
munch==2.5.0
nest-asyncio==1.5.6
networkx==3.1
nominatim==0.1
numpy==1.24.2
opencv-python==4.7.0.68
osmium==3.6.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pdf2image==1.16.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
pyproj @ file:///C:/Users/Administrator/Downloads/pyproj-3.3.1-cp38-cp38-win_amd64.whl#sha256=5dac03d4338a4c8bd0f69144c527474f517b4cbd7d2d8c532cd8937799723248
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
rapidfuzz==2.15.1
rasterio==1.3.6
requests==2.28.2
scikit-build==0.16.6
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Shapely @ file:///C:/Users/Administrator/Downloads/Shapely-1.8.2-cp38-cp38-win_amd64.whl#sha256=27af42a2f0374c8d1c0f62a8272ac28816c8e60f642e2934f46d0015e7b1c347
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wcwidth==0.2.6
xmltodict==0.13.0
xyzservices==2023.2.0
zipp==3.15.0
